SERVER-83441 Implement top-level fields pushdown from Filter into CollScan (#17683)
SERVER-83441 Implement top-level fields pushdown from Filter into
CollScan

Push top-level PathGet as a PhysicalScan projection.

Implemented via path analysis inside the implementor transport for
FilterNode. If there any fields to be pushed down, the transport creates
both a PhysicalScan node with the pushed fields, and FilterNode with
modified paths, where each top-level PathGet is removed.

The transformation is performed only for FilterNodes that are directly
on top of a ScanNode. In order to be efficient, this transformation
assumes that the FilterNode contains all the paths and path conditions,
that is, top-level conjunctions have not been split into a sequence of
FilterNode.

GitOrigin-RevId: 32cba0c99ac4d8e758068be6a724c0a6bc2d1ab0