SERVER-82881 Constructor for initializing BSONColumnBuilder from compressed binary (#17415)
SERVER-82881 Constructor for initializing BSONColumnBuilder from
compressed binary

Allows for efficient appending of more data in the scalar case to a
BSONColumn. This is orders of magnitude faster for larger BSONColumn
compared to the naive approach.

The end state of the BSONColumnBuilder is equivalent to one that
appended the contents of the binary but without calling finalize().

GitOrigin-RevId: e9dcfcc707f268fa1f55bd963ec8082e637250a6