SERVER-89184 Disable explain_find_trivially_false_predicates_in_agg_p… (#20976)
SERVER-89184 Disable
explain_find_trivially_false_predicates_in_agg_pipelines.js under tests
with random_migrations: true

Test configurations that set "random_migrations: true" cause the planner
to choose a different sharding plan for the lookup query in the test.

Explain for this query doesn't show any information about the sharded
part of the plan so it is not possible to verify whether it used EOF or
not.

The solution is to disable the test under those configurations that set
this flag.

GitOrigin-RevId: 12cc82e76349c5b953a2670c9def4905d5e414ad