SERVER-89673: Correct lifetime of BSONElement in getTopBottomNValueExprHelper (#21399)
SERVER-89673

Correct lifetime of BSONElement in getTopBottomNValueExprHelper()

`getTopBottomNValueExprHelper()` returned a `BSONElement` referencing
data owned by a `BSONObj` on the stack.
This was "sometimes" safe, as `Document::toBson()` may return a copy of
a pre-existing `BSONObj`, meaning the underlying buffer has other shared
owners, and can outlive the stack `BSONObj`. However, for non
`isTriviallyConvertible()` documents, a new buffer will be allocated for
the stack `BSONObj`, which will be the sole owner, destroying it when it
goes out of scope.

To prevent such issues in the future, this patch annotates
`BSONObj::getField()` with `[[lifetimebound]]`. This allows the compiler
to catch cases where the returned element clearly outlives the object,
and issue a warning. This allows some enforcement of the documented
contract of `BSONElement` which indicates it should not outlive the
parent `BSONObj`.

By making this a concrete requirement, use of `BSONElement` becomes much
less brittle; previously its safety often relied on un-enforced
expectation that the underlying buffer is owned elsewhere - which could
be violated by unrelated changes to buffer ownership.

GitOrigin-RevId: 4017b857ce804b6edd259ddb204eeb7ebf530d2e