SERVER-90935: Mount session before getting git file content (#22640)
# Issue

[SERVER-90935](https://jira.mongodb.org/browse/SERVER-90935), to fix
[BF-31482](https://jira.mongodb.org/browse/BF-31482).

# Description

The `Transport endpoint is not connected` error seems to occur after a
period of inactivity. I found similar reports on the web and some
evidence in our own repo that effectively remount the session to try to
avoid this.

I found no evidence in the logs that the `@retry` was really retrying 15
times, based on log timestamps. Rather than track down the subtleties
there, it seems like using a remount is a more probable solution.

# Testing

https://spruce.mongodb.com/version/66560d640498d900073e2f43/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

GitOrigin-RevId: ebc4cf08af19bb9bbd4f02cc67a7652b0e9cf7c2